Renaming files to use singular convention
